Merging changes...

+++ kotor Randomizer 2/KPaths.cs
@@ -32,9 +32,13 @@ namespace kotor_Randomizer_2
         public readonly string sounds;
         /// <summary> Path to the swkotor\TexturePacks directory. </summary>
         public readonly string TexturePacks;
+        /// <summary> Path to the dialog.tlk file within the swkotor directory </summary>
+        public readonly string dialog;
 
         /// <summary> Path to the RANDOMIZED.log file within the swkotor directory. </summary>
         public readonly string RANDOMIZED_LOG;
+        /// <summary> Filename of the log file indicating that the game has been randomized. </summary>
+        public const string RANDOMIZED_LOG_FILENAME = "RANDOMIZED.log";
 
         /// <summary> Path to the backup of the chitin.key file within the swkotor directory. </summary>
         public readonly string chitin_backup;
@@ -54,6 +58,8 @@ namespace kotor_Randomizer_2
         public readonly string sounds_backup;
         /// <summary> Path to the backup of the swkotor\TexturePacks directory. </summary>
         public readonly string TexturePacks_backup;
+        /// <summary> Path to the backup of the dialog.tlk file within the swkotor directory. </summary>
+        public readonly string dialog_backup;
 
         /// <summary>
         /// Constructs paths to the SW KotOR directory and subdirectories.
@@ -71,8 +77,9 @@ namespace kotor_Randomizer_2
             music = $"{swkotor_path}\\streammusic\\";
             sounds = $"{swkotor_path}\\streamsounds\\";
             TexturePacks = $"{swkotor_path}\\TexturePacks\\";
+            dialog = $"{swkotor_path}\\dialog.tlk";
 
-            RANDOMIZED_LOG = $"{swkotor_path}\\RANDOMIZED.log";
+            RANDOMIZED_LOG = $"{swkotor_path}\\{RANDOMIZED_LOG_FILENAME}";
 
             chitin_backup = $"{swkotor_path}\\chitin.key.bak";
             data_backup = $"{swkotor_path}\\data_bak\\";
@@ -83,6 +90,7 @@ namespace kotor_Randomizer_2
             music_backup = $"{swkotor_path}\\streammusic_bak\\";
             sounds_backup = $"{swkotor_path}\\streamsounds_bak\\";
             TexturePacks_backup = $"{swkotor_path}\\TexturePacks_bak\\";
+            dialog_backup = $"{swkotor_path}\\dialog.tlk.bak";
         }
 
         /// <summary> Returns a list of the current files in the swkotor base directory. </summary>
@@ -122,8 +130,7 @@ namespace kotor_Randomizer_2
         { get { return new DirectoryInfo(Override).GetFiles(); } }
 
         /// <summary> Returns a list of the current files in the swkotor\rims directory. </summary>
-        public FileInfo[] FilesInRims
-        { get { return new DirectoryInfo(rims).GetFiles(); } }
+        public FileInfo[] FilesInRims => Directory.Exists(rims) ? new DirectoryInfo(rims).GetFiles() : null;
 
         /// <summary> Returns a list of the current files in the swkotor\TexturePacks directory. </summary>
         public FileInfo[] FilesInTexturePacks
@@ -231,13 +238,25 @@ namespace kotor_Randomizer_2
         }
 
         /// <summary>
+        /// Creates a backup of the dialog file if it doesn't exist already.
+        /// </summary>
+        public void BackUpDialogFile()
+        {
+            if (!File.Exists(dialog_backup))
+            {
+                File.Copy(dialog, dialog_backup);
+            }
+        }
+
+        /// <summary>
         /// If the backup Modules directory exists, restore it to the active directory.
         /// </summary>
         public void RestoreModulesDirectory()
         {
             if (Directory.Exists(modules_backup))
             {
-                Directory.Delete(modules, true);
+                if (Directory.Exists(modules))
+                    Directory.Delete(modules, true);
                 Directory.Move(modules_backup, modules);
             }
         }
@@ -249,7 +268,8 @@ namespace kotor_Randomizer_2
         {
             if (Directory.Exists(lips_backup))
             {
-                Directory.Delete(lips, true);
+                if (Directory.Exists(lips))
+                    Directory.Delete(lips, true);
                 Directory.Move(lips_backup, lips);
             }
         }
@@ -261,7 +281,8 @@ namespace kotor_Randomizer_2
         {
             if (Directory.Exists(Override_backup))
             {
-                Directory.Delete(Override, true);
+                if (Directory.Exists(Override))
+                    Directory.Delete(Override, true);
                 Directory.Move(Override_backup, Override);
             }
         }
@@ -273,7 +294,8 @@ namespace kotor_Randomizer_2
         {
             if (Directory.Exists(music_backup))
             {
-                Directory.Delete(music, true);
+                if (Directory.Exists(music))
+                    Directory.Delete(music, true);
                 Directory.Move(music_backup, music);
             }
         }
@@ -285,7 +307,8 @@ namespace kotor_Randomizer_2
         {
             if (Directory.Exists(sounds_backup))
             {
-                Directory.Delete(sounds, true);
+                if (Directory.Exists(sounds))
+                    Directory.Delete(sounds, true);
                 Directory.Move(sounds_backup, sounds);
             }
         }
@@ -297,7 +320,8 @@ namespace kotor_Randomizer_2
         {
             if (Directory.Exists(TexturePacks_backup))
             {
-                Directory.Delete(TexturePacks, true);
+                if (Directory.Exists(TexturePacks))
+                    Directory.Delete(TexturePacks, true);
                 Directory.Move(TexturePacks_backup, TexturePacks);
             }
         }
@@ -309,8 +333,22 @@ namespace kotor_Randomizer_2
         {
             if (File.Exists(chitin_backup))
             {
-                File.Delete(chitin);
+                if (File.Exists(chitin))
+                    File.Delete(chitin);
                 File.Move(chitin_backup, chitin);
+            }
+        }
+
+        /// <summary>
+        /// If the backup dialog file exists, restore it to the active directory.
+        /// </summary>
+        public void RestoreDialogFile()
+        {
+            if (File.Exists(dialog_backup))
+            {
+                if (File.Exists(dialog))
+                    File.Delete(dialog);
+                File.Move(dialog_backup, dialog);
             }
         }
